Fixed sizet warnings (#100)
* don't cycle playlist if on "street fighter and capcom fighters" or "street fighter"

* lower errors to warnings in logs

* - fixed sort to be more supportive of sort types (monkofthefunk)
 - fixed playcount as an reloadableimage (monkofthfunk)
 - improve restart video fast scroll flicker (monkofthefunk)

* fixed reload playcount after launch for images

* linuxfix

* fix

* fixed high GPU due to marquee rendering thread on game launch

* revert changes

* fix size_t warnings

+++ RetroFE/Source/Collection/CollectionInfoBuilder.cpp
@@ -542,7 +542,7 @@ void CollectionInfoBuilder::loadPlaylistItems(CollectionInfo* info, std::map<std
         std::string basename = (std::string::npos == position) ? file : file.substr(0, position);
 
         std::string comparator = ".txt";
-        int start = file.length() - comparator.length();
+        size_t start = file.length() - comparator.length();
 
         if (start >= 0)
         {
@@ -924,7 +924,7 @@ void CollectionInfoBuilder::ImportRomDirectory(std::string path, CollectionInfo 
                 for(extensionsIt = extensions.begin(); extensionsIt != extensions.end(); ++extensionsIt)
                 {
                     std::string comparator = "." + *extensionsIt;
-                    int start = file.length() - comparator.length() + 1;
+                    size_t start = file.length() - comparator.length() + 1;
 
                     if (start >= 0)
                     {
