+++ RetroFE/Source/RetroFE.cpp
@@ -2263,7 +2263,7 @@ CollectionInfo *RetroFE::getCollection(std::string collectionName)
         std::string basename = (std::string::npos == position)? file : file.substr( 0, position );
 
         std::string comparator = ".sub";
-        int start = file.length( ) - comparator.length( );
+        size_t start = file.length( ) - comparator.length( );
 
         if ( start >= 0 )
         {
